+++ CHANGELOG.md
@@ -65,7 +65,7 @@
 + Added **Tooltip Tweaks Breakdown Hotkey** - When the keybind is pressed, show a breakdown of all fortune sources on a tool. (contributed by appable)
 + Added **Tooltip Tweaks Tooltip Format** - Show crop-specific farming fortune in tooltip. (contributed by appable)
 + Added command **/shcropspeedmeter** - Helps calculate the real farming fortune with the formula crops broken per block.
-
++ Added **Compost Low Notification** - Shows a notification as title when organic matter/fuel is low.
 
 ### Features from other Mods
 > *The following features are only there because I want them when testing SkyHanni features without other mods present.*